Add support for passing keyword arguments in an object to high state.
High state currently requires the state body to be an array of key/value pair objects. This a poor design when using JSON. Add support for passing keyword arguments in a object to salt states similar to salt modules. Convert the object to the expected array of key/value pairs.